+++ structs/requests.go
@@ -4,7 +4,7 @@ type (
 	CatalogItemRequest struct {
 		Title       string `json:"title"`
 		Description string `json:"description"`
-		Price       int    `json:"price"`
+		Price       float32    `json:"price"`
 		CoverUrl    string `json:"cover_url"`
 		Currency    string `json:"currency"`
 		// CatalogID   int    `json:"catalog_id"`
